chore: AuthPage => Auth로 폴더이름 변경 및 router 이름 변경
