+++ src/pages/Auth/Auth.module.scss
@@ -1,0 +1,45 @@
+@import "src/utils/scss/media";
+
+.template{
+  width: 100%;
+  min-height: 100vh;
+  display: flex;
+  -webkit-box-pack: center;
+  justify-content: center;
+  -webkit-box-align: center;
+  align-items: center;
+  text-align: center;
+  
+  @include media-breakpoint(mobile) {
+    text-align: center;
+    display: block;
+  }
+
+  &__content {
+    width: 390px;
+    margin: 80px auto 50px auto;
+
+    @include media-breakpoint(mobile) {
+      width: 360px;
+      margin: 80px auto;
+    }
+  }
+  &__logo {
+    display: inline-block;
+    margin-bottom: 40px;
+    
+    @include media-breakpoint(mobile) {
+      margin-bottom: 60px;
+      margin-right: 17px;
+    }
+    &__img{
+      width: 125px;
+      cursor: pointer;
+
+      @include media-breakpoint(mobile) {
+        width: 105px;
+      }
+    }
+  }
+
+}
