+++ src/pages/Auth/LoginPage/LoginPage.module.scss
@@ -1,19 +1,10 @@
 @import "src/utils/scss/media";
 
 .template{
-
-  &__content {
-    width: 390px;
-    margin: 80px auto 50px; 
-    display: flex;
-    flex-direction: column; 
-
-    @include media-breakpoint(mobile) {
-        width: 360px;
-        margin: 80px auto;
-    }
-  }
-    
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
 
   &__copyright {
     font-family: NanumSquareR, sans-serif;
