--- src/pages/AuthPage/AuthPage.module.scss
@@ -1,21 +1,0 @@
-@import "src/utils/scss/media";
-
-.template{
-  width: 100%;
-  min-height: 100vh;
-  display: flex;
-  -webkit-box-pack: center;
-  justify-content: center;
-  -webkit-box-align: center;
-  align-items: center;
-  flex-direction: column;
-
-  &__logo {
-    width: 125px;
-    cursor: pointer;
-    
-    @include media-breakpoint(mobile) {
-      width: 105px;
-    }
-  }
-}
